175 added, 163 modified, 26 deleted C:\save\modding\ervin_mod_1

+++ src/main/java/com/babcsany/minecraft/ervin_mod_1/block/FallingZurBlock.java
@@ -41,7 +41,7 @@ public class FallingZurBlock extends Block {
 
    public void tick(BlockState state, ServerWorld worldIn, BlockPos pos, Random rand) {
       if (worldIn.isAirBlock(pos.down()) || canFallThrough(worldIn.getBlockState(pos.down())) && pos.getY() >= 0) {
-         FallingBlockEntity fallingblockentity = new FallingBlockEntity(worldIn, (double)pos.getX() + 0.5D, (double)pos.getY(), (double)pos.getZ() + 0.5D, worldIn.getBlockState(pos));
+         FallingBlockEntity fallingblockentity = new FallingBlockEntity(worldIn, (double)pos.getX() + 0.5D, pos.getY(), (double)pos.getZ() + 0.5D, worldIn.getBlockState(pos));
          this.onStartFalling(fallingblockentity);
          worldIn.addEntity(fallingblockentity);
       }
